authorization: Authorization middleware added and routes has been protected

+++ day70-b43-backend/controllers/tasks.controller.js
@@ -3,7 +3,10 @@ const taskModel = require("../models/tasks.model");
 // GET
 const getallTask = async (req, res) => {
     try {
-        const tasks = await taskModel.find();
+        const tasks = await taskModel.find({
+            user: req.userId
+        }).sort({ createdAt: -1 });
+
         res.status(200).json(tasks);
     } catch (err) {
         res.status(500).json({ message: "Server error, please try again after sometimes" });
@@ -16,7 +19,9 @@ const createTask = async (req, res) => {
     try {
         const { title, description, status } = req.body;
 
-        const newTask = await taskModel.create({ title, description, status });
+        const newTask = await taskModel.create(
+            { title, description, status, user: req.userId }
+        );
 
         res.status(201).json(newTask);
 
@@ -27,21 +32,42 @@ const createTask = async (req, res) => {
 
 // PUT
 const updateTask = async (req, res) => {
-    try{
-        const updatedTask = await taskModel.findByIdAndUpdate(req.params.id, req.body, {new: true});
-        res.status(200).json({message: "Task updated"});
-    }catch(err){
-        res.status(500).json({message: "Server error, please try again after sometimes"});
+    try {
+        // ownership check
+        const task = await taskModel.findOne({
+            _id: req.params.id,
+            user: req.userId
+        });
+
+        if (!task) {
+            return res.status(401).json({ message: "Unauthorized: Access denied" });
+        }
+
+        const updatedTask = await taskModel.findByIdAndUpdate(req.params.id, req.body, { new: true });
+        res.status(200).json({ message: "Task updated" });
+    } catch (err) {
+        res.status(500).json({ message: "Server error, please try again after sometimes" });
     }
 };
 
 // DELETE
 const deleteTask = async (req, res) => {
-    try{
+    try {
+
+        // ownership check
+        const task = await taskModel.findOne({
+            _id: req.params.id,
+            user: req.userId
+        });
+
+        if (!task) {
+            return res.status(401).json({ message: "Unauthorized: Access denied" });
+        }
+
         await taskModel.findByIdAndDelete(req.params.id);
-        res.status(200).json({message: "Task deleted"});
-    }catch(err){
-        res.status(500).json({message: "Server error, please try again after sometimes"});
+        res.status(200).json({ message: "Task deleted" });
+    } catch (err) {
+        res.status(500).json({ message: "Server error, please try again after sometimes" });
     }
 };
